- disable unsafe reflection usage by default
