+++ src/main/java/mediathek/Main.java
@@ -319,6 +319,7 @@ public class Main {
      * @param args the command line arguments
      */
     public static void main(final String... args) {
+        Functions.disableAccessWarnings();
         setupEnvironmentProperties();
 
         if (GraphicsEnvironment.isHeadless()) {
